services/warehouse/warehouse_app/views.py
from django.core.cache import cache
from .models import Item
from .serializers import ItemSerializer
from rest_framework.views import APIView
from rest_framework.response import Response


class ItemListCreate(APIView):

    def get(self, request):
        # Try to get the items from the cache first
        items = cache.get("items_key", version=1)

        if not items:  # If cache miss, fetch from DB and set to cache
            items = Item.objects.all()
            cache.set("items_key", items, version=1)

        serializer = ItemSerializer(items, many=True)
        return Response(serializer.data)

    def post(self, request):
        serializer = ItemSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()

            # Invalidate or update the cache when a new item is added
            items = Item.objects.all()
            cache.set("items_key", items, version=2)

            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)


class ItemDetail(APIView):
    def get_object(self, pk):
        return get_object_or_404(Item, pk=pk)

    def get(self, request, pk):
        item = self.get_object(pk)
        serializer = ItemSerializer(item)
        return Response(serializer.data)

    def put(self, request, pk):
        item = self.get_object(pk)
        serializer = ItemSerializer(item, data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
