services/warehouse/requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework
django-redis
django-environ
psycopg2-binary